fpc_imp: L-T-Y: Allocate and use global buffer for normal command
As also stated by the TODO comment, allocating and freeing the
buffer again and again is useless and only makes us waste
precious CPU cycles and bandwidth because of the continuously
called memsets.

Because that buffer is always of a known 0x40 length, implement
a global ION buffer and use it for the normal command.

Change-Id: Ife5a5e7aecc9749ec4214a8e25d96c0fc538cdb7