fingerprint: store empty user db with GID once it is created
When no user db is found, an empty one would be created.
However, the current logic does not set the correct GID along, and instead just stores the empty db and returns.
This patch makes the storing of the empty user db happen only after GID is set.
Also, improve the logging a bit to distinguish between the two possible cases of either having a previous db or creating a new one.

Change-Id: I67e3775ad051b691312d7c13c3f3875a385e0dba